Refactored the constructor resolution code to live in extension methods
