+++ src/Retro.FastInject/Generation/ResolutionExtensions.cs
@@ -1,4 +1,14 @@
-﻿using Retro.FastInject.Model.Manifest;
+﻿using System;
+using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Diagnostics.CodeAnalysis;
+using Microsoft.CodeAnalysis;
+using Retro.FastInject.Model.Manifest;
+using System.Linq;
+using System.Text;
+using Microsoft.Extensions.DependencyInjection;
+using Retro.FastInject.Annotations;
+using Retro.FastInject.Utils;
 
 namespace Retro.FastInject.Generation;
 
@@ -7,6 +17,266 @@ namespace Retro.FastInject.Generation;
 /// used in the dependency injection service hierarchy.
 /// </summary>
 public static class ResolutionExtensions {
+  /// <summary>
+  /// Checks if all dependencies in the constructor of the specified service registration can be resolved and records the resolution.
+  /// </summary>
+  /// <param name="declaration">The service registration representing the type and its associated symbol to check.</param>
+  /// <param name="compilation">The Roslyn compilation providing semantic analysis for the type and its dependencies.</param>
+  /// <exception cref="InvalidOperationException">
+  /// Thrown when dependencies cannot be resolved, or when the type has multiple public constructors, or if the type is not a named type.
+  /// </exception>
+  public static void CheckConstructorDependencies(this ServiceManifest serviceManifest, 
+                                           ServiceRegistration declaration, Compilation compilation) {
+    var type = declaration.Type;
+    if (type is not INamedTypeSymbol namedTypeSymbol) {
+      throw new InvalidOperationException($"Type '{type.ToDisplayString()}' is not a named type.");
+    }
+
+    var constructor = declaration.AssociatedSymbol switch {
+        IMethodSymbol method => method.ValidateFactoryMethod(),
+        null => namedTypeSymbol.GetValidConstructor(),
+        _ => null
+    };
+    if (constructor is null) {
+      return;
+    }
+
+    var missingDependencies = new List<string>();
+
+    // Create constructor resolution record
+    var constructorResolution = new ConstructorResolution {
+        Constructor = constructor,
+        Type = type
+    };
+
+    foreach (var parameter in constructor.Parameters) {
+      serviceManifest.ResolveParameterDependencies(parameter, constructorResolution, compilation, missingDependencies);
+    }
+
+    // Store the constructor resolution
+    serviceManifest.AddConstructorResolution(constructorResolution);
+
+    if (missingDependencies.Count > 0) {
+      throw new InvalidOperationException(
+          $"Cannot resolve the following dependencies for type '{type.ToDisplayString()}':\n" +
+          $"- {string.Join("\n- ", missingDependencies)}");
+    }
+  }
+  
+  private static IMethodSymbol? GetValidConstructor(this INamedTypeSymbol type) {
+    var publicConstructors = type.Constructors
+        .Where(c => c.DeclaredAccessibility == Accessibility.Public)
+        .ToArray();
+
+    var explicitConstructors = publicConstructors
+        .Where(x => !x.IsImplicitlyDeclared)
+        .ToArray();
+
+    if (explicitConstructors.Length > 0) {
+      return explicitConstructors.Length switch {
+          > 1 => throw new InvalidOperationException(
+              $"Type '{type.ToDisplayString()}' has multiple public constructors. Only one public constructor is allowed for dependency injection."),
+          _ => explicitConstructors[0]
+      };
+    }
+
+    // Fallback to the first public constructor if there are no explicit constructors
+    return publicConstructors.FirstOrDefault();
+  }
+  
+  private static IMethodSymbol ValidateFactoryMethod(this IMethodSymbol methodSymbol) {
+    if (methodSymbol.ReturnsVoid || methodSymbol.ReturnType is not INamedTypeSymbol) {
+      throw new InvalidOperationException(
+          $"Factory method '{methodSymbol.ToDisplayString()}' must return a named non-void type.");
+    }
+
+    return methodSymbol;
+  }
+  
+  private static void ResolveParameterDependencies(this ServiceManifest serviceManifest, IParameterSymbol parameter, ConstructorResolution constructorResolution,
+                                            Compilation compilation,
+                                            List<string> missingDependencies) {
+    var (isNullable, paramType) = parameter.Type.CheckIfNullable();
+    
+    // Create parameter resolution
+    var parameterResolution = new ParameterResolution {
+        Parameter = parameter,
+        ParameterType = paramType
+    };
+
+    // Check for FromKeyedServices attribute
+    var fromKeyedServicesAttr = parameter.GetAttributes()
+        .FirstOrDefault(a => a.IsOfAttributeType<FromKeyedServicesAttribute>());
+
+    string? keyName = null;
+    if (fromKeyedServicesAttr is { ConstructorArguments.Length: > 0 }) {
+      keyName = fromKeyedServicesAttr.ConstructorArguments[0].Value?.ToString();
+    }
+
+    parameterResolution.Key = keyName;
+  
+    var canResolve = serviceManifest.CanResolve(keyName, paramType, parameterResolution, compilation, 
+                                                out var selectedService);
+
+    parameterResolution.SelectedService = selectedService;
+    parameterResolution.DefaultValue = parameter.GetDefaultValueString();
+    constructorResolution.Parameters.Add(parameterResolution);
+
+    if (canResolve || isNullable || parameterResolution.DefaultValue is not null) return;
+      
+    // Add the missing dependency to the list with detailed information
+    var dependency = new StringBuilder(paramType.ToDisplayString());
+    if (keyName != null) {
+      dependency.Append($" with key '{keyName}'");
+    }
+
+    // Add more specific error information
+    if (parameterResolution.HasNoDeclaration) {
+      dependency.Append(" (No service declaration found)");
+    } else if (parameterResolution is { HasMultipleRegistrations: true, MultipleServices.Count: > 0 }) {
+      dependency.Append($" (Multiple registrations found: {parameterResolution.MultipleServices.Count})");
+      // Optionally add details about the multiple registrations
+      foreach (var service in parameterResolution.MultipleServices) {
+        var implType = service.ImplementationType?.ToDisplayString() ?? service.Type.ToDisplayString();
+        dependency.Append($"\n  -- {implType}" + (service.Key != null ? $" with key '{service.Key}'" : ""));
+      }
+    }
+  
+    missingDependencies.Add(dependency.ToString());
+  }
+  
+  private static bool CanResolve(this ServiceManifest serviceManifest, 
+                                 string? keyName, 
+                                 ITypeSymbol paramType, 
+                                 ParameterResolution parameterResolution, 
+                                 Compilation compilation, 
+                                 out ServiceRegistration? selectedService) {
+    // Check if the dependency can be resolved
+    var canResolve = false;
+    selectedService = null;
+
+    if (serviceManifest.TryGetServices(paramType, out var registrations)) {
+      try {
+        selectedService = registrations.SingleOrDefault(r => keyName is null || r.Key == keyName);
+      } catch (InvalidOperationException) {
+        parameterResolution.HasMultipleRegistrations = true;
+        parameterResolution.MultipleServices = registrations.ToList();
+        selectedService = null;
+      }
+
+      if (selectedService is not null) {
+        selectedService = serviceManifest.ResolveConcreteType(selectedService);
+      }
+      
+      canResolve = selectedService != null;
+    } else {
+      // Check if the type is a collection type
+      if (keyName is null && paramType is INamedTypeSymbol { IsGenericType: true } namedType 
+                          && serviceManifest.CanResolveGenericType(namedType, compilation, parameterResolution,
+                                                   keyName, ref selectedService)) {
+        return true;
+      }
+      parameterResolution.HasNoDeclaration = true;
+    }
+
+    return canResolve;
+  }
+
+  private static bool CanResolveGenericType(this ServiceManifest serviceManifest, INamedTypeSymbol namedType, Compilation compilation, 
+                                           ParameterResolution targetParameter,
+                                           string? keyName,
+                                           ref ServiceRegistration? selectedService) {
+    var genericTypeName = namedType.ConstructedFrom.ToDisplayString();
+    if (genericTypeName is "System.Collections.Generic.IEnumerable<T>" or
+        "System.Collections.Generic.IReadOnlyCollection<T>" or
+        "System.Collections.Generic.IReadOnlyList<T>" or
+        "System.Collections.Immutable.ImmutableArray<T>") {
+      return serviceManifest.TryResolveServiceCollection(namedType, compilation, targetParameter.Parameter, 
+                                                         out selectedService);
+    }
+
+    if (!serviceManifest.TryGetServices(namedType.ConstructedFrom, out var registrations)) {
+      return false;
+    }
+
+    ServiceRegistration? unboundService = null;
+    try {
+      unboundService = registrations.SingleOrDefault(r => keyName is null || r.Key == keyName);
+    } catch (InvalidOperationException) {
+      targetParameter.HasMultipleRegistrations = true;
+      targetParameter.MultipleServices = registrations.ToList();
+    }
+
+    if (unboundService is null) return false;
+
+    var concreteUnbound = serviceManifest.ResolveConcreteType(unboundService);
+    var implementationType = concreteUnbound.Type.GetInstantiatedGeneric(namedType.TypeArguments.ToArray());
+      
+    selectedService = serviceManifest.AddService(implementationType, unboundService.Lifetime, key: unboundService.Key, associatedSymbol: unboundService.AssociatedSymbol);
+    foreach (var superclass in implementationType.GetAllSuperclasses()
+                 .Where(x => !x.IsSpecialInjectableType() && !x.Equals(implementationType, SymbolEqualityComparer.Default))) {
+      serviceManifest.AddService(superclass, unboundService.Lifetime, implementationType, unboundService.AssociatedSymbol, unboundService.Key);
+      
+      // We need to add this type to any collection declarations that may already exist
+      var immutableArrayType = typeof(ImmutableArray<>).GetInstantiatedGeneric(compilation, superclass);
+      if (!serviceManifest.TryGetServices(immutableArrayType, out var immutableArrayServices)) continue;
+
+      foreach (var service in immutableArrayServices) {
+        service.CollectedServices?.Add(selectedService);
+      }
+    }
+    serviceManifest.CheckConstructorDependencies(selectedService, compilation);
+      
+    return true;
+  }
+  private static bool TryResolveServiceCollection(this ServiceManifest serviceManifest, INamedTypeSymbol namedType, Compilation compilation, IParameterSymbol targetParameter, 
+                                           [NotNullWhen(true)] out ServiceRegistration? selectedService) {
+    var elementType = namedType.TypeArguments[0];
+    if (!serviceManifest.TryGetServices(elementType, out var elementServices)) {
+      elementServices = [];
+    }
+
+    var requireNonEmptyAttribute = targetParameter.GetAttribute<RequireNonEmptyAttribute>();
+    if (requireNonEmptyAttribute is not null && elementServices.Count == 0) {
+      selectedService = null;
+      return false;
+    }
+
+    var immutableArrayType = typeof(ImmutableArray<>).GetInstantiatedGeneric(compilation, elementType);
+    serviceManifest.AddService(immutableArrayType, ServiceScope.Transient);
+    var readOnlyListType = typeof(IReadOnlyList<>).GetInstantiatedGeneric(compilation, elementType);
+    serviceManifest.AddService(readOnlyListType, ServiceScope.Transient, immutableArrayType);
+    var readOnlyCollectionType = typeof(IReadOnlyCollection<>).GetInstantiatedGeneric(compilation, elementType);
+    serviceManifest.AddService(readOnlyCollectionType, ServiceScope.Transient, immutableArrayType);
+    var enumerableType = typeof(IEnumerable<>).GetInstantiatedGeneric(compilation, elementType);
+    selectedService = serviceManifest.AddService(enumerableType, ServiceScope.Transient, immutableArrayType,
+                                                 collectedServices: elementServices
+                                                     .Select(serviceManifest.ResolveConcreteType)
+                                                     .ToList());
+    return true;
+  }
+
+  private static ServiceRegistration ResolveConcreteType(this ServiceManifest serviceManifest, 
+                                                         ServiceRegistration declaration) {
+    if (declaration.ImplementationType is null || !serviceManifest.TryGetServices(declaration.ImplementationType, 
+                                                                               out var possibleImpls)) {
+      return declaration;
+    }
+    
+    try {
+      var implService = possibleImpls
+          .SingleOrDefault(x => declaration.Key is null || declaration.Key == x.Key);
+      if (implService is not null) {
+        return implService;
+      }
+    } catch (InvalidOperationException) {
+      // In this case swallow the exception and keep the abstract service
+      // as the selected service.
+    }
+    
+    return declaration;
+  }
+  
   /// <summary>
   /// Retrieves the default value of the argument for the provided parameter resolution.
   /// </summary>
