+++ test/Retro.FastInject.Tests/ServiceHierarchy/ServiceManifestTest.cs
@@ -6,6 +6,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using NUnit.Framework;
 using Retro.FastInject.Annotations;
+using Retro.FastInject.Generation;
 using Retro.FastInject.Model.Manifest;
 using Retro.FastInject.Tests.Utils;
